requirements.txt
streamlit==1.32.0
google-generativeai==0.3.2
python-dotenv==1.0.0
pinecone-client==3.0.2
PyPDF2==3.0.1
langchain==0.1.9
sentence-transformers==2.5.1 